courses-api-main/requirements.txt
aniso8601==9.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
google-api-core==1.31.2
google-auth==1.35.0
google-cloud-dialogflow==2.7.1
google-cloud-texttospeech==2.5.2
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
